End test immediately when all streams finish
The caller expects a test's update channel to close when the test is
complete. Previously, even when all streams had finished, the test was
not "finished" and the update channel did not close until after a timer
had gone off indicating that the server had not ended the test.
Correctly detect when the server has ended the test and close the update
channel immediately.